Correct the location of dics and files to follow the convention.
