+++ examples/grad/04-spin_flip_tddft_grad.py
@@ -7,12 +7,12 @@ the gradient module.
 import numpy as np
 from pyscf import lib, gto, dft
 from pyscf import sftda
-from pyscf.sftda_grad import tduks_sf # this import is necessary.
+from pyscf.grad import tduks_sf # this import is necessary.
 try:
     import mcfun # mcfun>=0.2.5 must be used.
 except ImportError:
     mcfun = None
-    
+
 mol = gto.Mole()
 mol.verbose = 6
 mol.output = None
